requirements.txt
PyQt5==5.15.9
PyQtWebEngine==5.15.6
pyautogui==0.9.54
keyboard==0.13.5
pyyaml==6.0.1
cryptography==41.0.5
requests==2.31.0
python-dotenv==1.0.1
SQLAlchemy==2.0.27
aiohttp==3.9.1
websockets==12.0
qasync==0.24.0
pillow==10.1.0
numpy==1.26.2
opencv-python==4.8.1.78
pyinstaller==6.2.0
bcrypt==4.0.1
pyjwt==2.8.0
python-multipart==0.0.6
aiosqlite==0.19.0
pywin32==306; platform_system == "Windows"
psutil==5.9.7
pygame==2.5.2
psycopg2-binary==2.9.9
pytest==8.3.5
pytest-qt==4.4.0
black==23.11.0
flake8==6.1.0
mypy==1.7.0
fastapi==0.104.1
uvicorn==0.24.0
python-jose==3.3.0
passlib==1.7.4
pydantic==2.5.2 